+++ Magic/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fix an unintentional behavioral change in undo animations that was introduced in 7.2
  - Add support for falling block projectiles
  - Fix issues using /reload command (still don't use it, though...)
+ - Add /mage clear command for clearing player inventories of magic items
 
 # 7.3.7